rename OfferCardSquare to OfferCardNarrow, rename Bookmark to BookmarkButton
