+++ src/components/offer-card/offer-card-narrow/OfferCardNarrow.styles.ts
